test: minor restructuring + add basic test for inline entrypoint
